fixing input error
Real quick, noticed that KeyEvent was also being returned as part of the input.  Since it is not a function declared as part of Game.input, it cannot be returned and was causing an error with the gameplay screen.  Figured I would change it quick so that we don't get confused later.